Allow inheritance in testing.inject_backend_tests
